# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wisata Indonesia API - Tourism Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Province {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  createdAt: string;
}

export interface CreateProvinceInput {
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
}

export interface City {
  id: number;
  provinceId: number;
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  province?: Province | null;
  createdAt: string;
}

export interface CreateCityInput {
  provinceId: number;
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
}

export interface Destination {
  id: number;
  cityId: number;
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  rating?: number | null;
  city?: City | null;
  createdAt: string;
}

export interface CreateDestinationInput {
  cityId: number;
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  rating?: number | null;
}

export interface TourPackage {
  id: number;
  title: string;
  slug: string;
  description?: string | null;
  price: number;
  duration: number;
  maxParticipants?: number | null;
  imageUrl?: string | null;
  images: string[];
  includes: string[];
  excludes: string[];
  itinerary?: string | null;
  provinceId?: number | null;
  cityId?: number | null;
  isFeatured: boolean;
  isActive: boolean;
  rating?: number | null;
  totalReviews: number;
  province?: Province | null;
  city?: City | null;
  createdAt: string;
}

export interface CreatePackageInput {
  title: string;
  slug: string;
  description?: string | null;
  price: number;
  duration: number;
  maxParticipants?: number | null;
  imageUrl?: string | null;
  images?: string[];
  includes?: string[];
  excludes?: string[];
  itinerary?: string | null;
  provinceId?: number | null;
  cityId?: number | null;
  isFeatured: boolean;
  isActive: boolean;
}

export interface Vehicle {
  id: number;
  name: string;
  type: string;
  brand?: string | null;
  capacity: number;
  pricePerDay: number;
  imageUrl?: string | null;
  images: string[];
  features: string[];
  isAvailable: boolean;
  description?: string | null;
  createdAt: string;
}

export interface CreateVehicleInput {
  name: string;
  type: string;
  brand?: string | null;
  capacity: number;
  pricePerDay: number;
  imageUrl?: string | null;
  images?: string[];
  features?: string[];
  isAvailable: boolean;
  description?: string | null;
}

export interface Booking {
  id: number;
  type: string;
  status: string;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  packageId?: number | null;
  vehicleId?: number | null;
  startDate: string;
  endDate?: string | null;
  participants?: number | null;
  totalPrice: number;
  notes?: string | null;
  package?: TourPackage | null;
  vehicle?: Vehicle | null;
  createdAt: string;
}

export interface CreateBookingInput {
  type: string;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  packageId?: number | null;
  vehicleId?: number | null;
  startDate: string;
  endDate?: string | null;
  participants?: number | null;
  totalPrice: number;
  notes?: string | null;
}

export interface UpdateBookingInput {
  status: string;
  notes?: string | null;
}

export interface GalleryItem {
  id: number;
  title: string;
  imageUrl: string;
  category?: string | null;
  destinationId?: number | null;
  createdAt: string;
}

export interface CreateGalleryItemInput {
  title: string;
  imageUrl: string;
  category?: string | null;
  destinationId?: number | null;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt?: string | null;
  content: string;
  imageUrl?: string | null;
  category?: string | null;
  tags: string[];
  authorName: string;
  isPublished: boolean;
  publishedAt?: string | null;
  createdAt: string;
}

export interface CreateBlogPostInput {
  title: string;
  slug: string;
  excerpt?: string | null;
  content: string;
  imageUrl?: string | null;
  category?: string | null;
  tags?: string[];
  authorName: string;
  isPublished: boolean;
}

export interface Testimonial {
  id: number;
  customerName: string;
  customerAvatar?: string | null;
  rating: number;
  comment: string;
  packageId?: number | null;
  isApproved: boolean;
  createdAt: string;
}

export interface CreateTestimonialInput {
  customerName: string;
  customerAvatar?: string | null;
  rating: number;
  comment: string;
  packageId?: number | null;
}

export interface UpdateTestimonialInput {
  isApproved: boolean;
  comment?: string | null;
}

export interface Banner {
  id: number;
  title: string;
  subtitle?: string | null;
  imageUrl: string;
  ctaText?: string | null;
  ctaLink?: string | null;
  isActive: boolean;
  order: number;
  createdAt: string;
}

export interface CreateBannerInput {
  title: string;
  subtitle?: string | null;
  imageUrl: string;
  ctaText?: string | null;
  ctaLink?: string | null;
  isActive: boolean;
  order: number;
}

export interface AdminStats {
  totalBookings: number;
  activeBookings: number;
  totalRevenue: number;
  totalPackages: number;
  totalVehicles: number;
  totalCustomTrips: number;
  recentBookings: Booking[];
}

export interface CustomTrip {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  destination: string;
  startDate: string;
  endDate?: string | null;
  participants: number;
  budget?: number | null;
  requirements?: string | null;
  status: string;
  createdAt: string;
}

export interface CreateCustomTripInput {
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  destination: string;
  startDate: string;
  endDate?: string | null;
  participants: number;
  budget?: number | null;
  requirements?: string | null;
}

export interface UpdateCustomTripInput {
  status: string;
}

export type GetCitiesParams = {
  provinceId?: number;
};

export type GetDestinationsParams = {
  cityId?: number;
  provinceId?: number;
};

export type GetPackagesParams = {
  provinceId?: number;
  cityId?: number;
  minPrice?: number;
  maxPrice?: number;
  duration?: number;
  featured?: boolean;
};

export type GetVehiclesParams = {
  type?: string;
  available?: boolean;
};

export type GetBookingsParams = {
  status?: string;
  type?: string;
};

export type GetGalleryParams = {
  category?: string;
};

export type GetBlogPostsParams = {
  category?: string;
  tag?: string;
  published?: boolean;
};

export type GetTestimonialsParams = {
  approved?: boolean;
};

export type GetBannersParams = {
  active?: boolean;
};