# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Wisata Indonesia API - Tourism Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /provinces:
    get:
      operationId: getProvinces
      tags: [provinces]
      summary: List all provinces
      responses:
        "200":
          description: List of provinces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Province"
    post:
      operationId: createProvince
      tags: [provinces]
      summary: Create a province
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProvinceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Province"
  /provinces/{id}:
    get:
      operationId: getProvince
      tags: [provinces]
      summary: Get a province
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Province
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Province"
    put:
      operationId: updateProvince
      tags: [provinces]
      summary: Update a province
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProvinceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Province"
    delete:
      operationId: deleteProvince
      tags: [provinces]
      summary: Delete a province
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /destinations:
    get:
      operationId: getDestinations
      tags: [destinations]
      summary: List all destinations
      parameters:
        - name: cityId
          in: query
          schema:
            type: integer
        - name: provinceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"
    post:
      operationId: createDestination
      tags: [destinations]
      summary: Create a destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDestinationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
  /destinations/{id}:
    get:
      operationId: getDestination
      tags: [destinations]
      summary: Get a destination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
    put:
      operationId: updateDestination
      tags: [destinations]
      summary: Update a destination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDestinationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
    delete:
      operationId: deleteDestination
      tags: [destinations]
      summary: Delete a destination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /packages:
    get:
      operationId: getPackages
      tags: [packages]
      summary: List tour packages
      parameters:
        - name: provinceId
          in: query
          schema:
            type: integer
        - name: cityId
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: duration
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TourPackage"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a tour package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourPackage"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get a tour package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tour package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourPackage"
    put:
      operationId: updatePackage
      tags: [packages]
      summary: Update a tour package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourPackage"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a tour package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /gallery:
    get:
      operationId: getGallery
      tags: [gallery]
      summary: List gallery items
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Create a gallery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Province:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - createdAt
    CreateProvinceInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required:
        - name
        - slug
    City:
      type: object
      properties:
        id:
          type: integer
        provinceId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        province:
          $ref: "#/components/schemas/Province"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - provinceId
        - name
        - slug
        - createdAt
...
    Destination:
      type: object
      properties:
        id:
          type: integer
        cityId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        city:
          $ref: "#/components/schemas/City"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - cityId
        - name
        - slug
        - createdAt
    CreateDestinationInput:
      type: object
      properties:
        cityId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
      required:
        - cityId
        - name
        - slug
    TourPackage:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        duration:
          type: integer
        maxParticipants:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        includes:
          type: array
          items:
            type: string
        excludes:
          type: array
          items:
            type: string
        itinerary:
          type: string
          nullable: true
        provinceId:
          type: integer
          nullable: true
        cityId:
          type: integer
          nullable: true
        isFeatured:
          type: boolean
        isActive:
          type: boolean
        rating:
          type: number
          nullable: true
        totalReviews:
          type: integer
        province:
          $ref: "#/components/schemas/Province"
          nullable: true
        city:
          $ref: "#/components/schemas/City"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - slug
        - price
        - duration
        - images
        - includes
        - excludes
        - isFeatured
        - isActive
        - totalReviews
        - createdAt
    CreatePackageInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        duration:
          type: integer
        maxParticipants:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        includes:
          type: array
          items:
            type: string
        excludes:
          type: array
          items:
            type: string
        itinerary:
          type: string
          nullable: true
        provinceId:
          type: integer
          nullable: true
        cityId:
          type: integer
          nullable: true
        isFeatured:
          type: boolean
        isActive:
          type: boolean
      required:
        - title
        - slug
        - price
        - duration
        - isFeatured
        - isActive
...
    GalleryItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        category:
          type: string
          nullable: true
        destinationId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - imageUrl
        - createdAt
    CreateGalleryItemInput:
      type: object
      properties:
        title:
          type: string
        imageUrl:
          type: string
        category:
          type: string
          nullable: true
        destinationId:
          type: integer
          nullable: true
      required:
        - title
        - imageUrl
...